+++ src/main/java/org/springframework/samples/nt4h/card/enemy/EnemyService.java
@@ -5,12 +5,14 @@ import org.springframework.samples.nt4h.action.Action;
 import org.springframework.samples.nt4h.action.InflictWounds;
 import org.springframework.samples.nt4h.action.RemoveCardForEnemyAttack;
 import org.springframework.samples.nt4h.game.Game;
+import org.springframework.samples.nt4h.player.Player;
 import org.springframework.samples.nt4h.player.PlayerService;
 import org.springframework.security.acls.model.NotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -80,28 +82,21 @@ public class EnemyService {
     }
     //Obetener el daño total de los enemigos en batalla
 
-    public void attackEnemyToActualPlayer(Game game) {
-        //Lo que hace el método :)
-        //Cambiar mazo de habilidad
-        //Recibir herida o no
-        Integer damage = 0; //repito el bucle porque no se me he encariñado de el
-        if (game.getActualOrcs().size() != 0) {  // Si hay enemigos en el campo si no pues no recibe daño el heroe
-            for (int i = 0; i <= game.getActualOrcs().size(); i++) {  //calculamos el daño total inflijido
-                damage = damage + game.getActualOrcs().get(i).getActualHealth();
+    public Integer attackEnemyToActualPlayer(Game game) {
+        Player currentPlayer = game.getCurrentPlayer();
+        if (game.getActualOrcs().isEmpty()) return 0;
+        // Si hay enemigos en el campo si no pues no recibe daño el heroe
+        int damage = game.getActualOrcs().stream().mapToInt(EnemyInGame::getActualHealth).sum();
+        if (currentPlayer.getInDeck().size() <= damage) { //si el daño es mayor o igual a la cantidad de cartass quue tengo pues recibo la herida
+            new InflictWounds(currentPlayer).executeAction();  //Recibe herida
+            if (Objects.equals(currentPlayer.getWounds(), currentPlayer.getHealth())) {
+                game.getPlayers().remove(currentPlayer);  // de momento sales de la partida, mas adelante cambia a vista espectador
+                playerService.getOutGame(currentPlayer, game);
             }
-            if (game.currentPlayer.getInDeck().size() <= damage) { //si el daño es mayor o igual a la cantidad de cartass quue tengo pues recibo la herida
-                Action DamageWithWounds = new InflictWounds(game.currentPlayer);  //Recibe herida
-                DamageWithWounds.executeAction();
-                if(game.currentPlayer.getWounds()==game.currentPlayer.getHeroes().get(0).getActualHealth()){
-                    game.getPlayers().remove(game.currentPlayer);  // de momento sales de la partida, mas adelante cambia a vista espectador
-                    playerService.getOutGame(game.currentPlayer, game);
-                }
-            } else {
-                Action DamageWithOUTWounds = new RemoveCardForEnemyAttack(game.currentPlayer, damage); //no recibe herida
-                DamageWithOUTWounds.executeAction();
-            }
+        } else {
+            new RemoveCardForEnemyAttack(currentPlayer, damage).executeAction(); //no recibe herida
         }
-        playerService.savePlayer(game.currentPlayer);
+        return damage;
     }
 
 }
